fix(linting): remove false positive linting
Eslint has had some false positives in the linting, where it marks the
arguments for functions passed as arguements as unused. This has been
the case in the Typescript definitions of the functions that are passed.
With this small fix, we ignore this case, since it is clearly a false
positive.
For reference:
https://github.com/eslint/eslint/issues/8661#issuecomment-304531311